# app/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API with Swagger",
        "description": "API documentation for Laravel using Swagger",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Display a listing of products",
                "operationId": "a31062dc8c5c8351b96c47fbe4f0534e",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by product name (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "Filter by brand ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by product status (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at_start",
                        "in": "query",
                        "description": "Filter by creation date start (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "created_at_end",
                        "in": "query",
                        "description": "Filter by creation date end (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/productPagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No products found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No products found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a newly created product",
                "operationId": "55c957b76252aec7cefa0f9fcb5677fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/storeProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Invalid request!"
                    },
                    "500": {
                        "description": "Internal Error!"
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Display the specified product",
                "operationId": "275c4d703cdf48428bd7d6542154af8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found!"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update the specified product",
                "operationId": "d23c5cabb1aa9757915f62f375f44d51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updateProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found!"
                    },
                    "422": {
                        "description": "Unprocessable Entity!"
                    },
                    "500": {
                        "description": "Internal Error!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Remove the specified product",
                "operationId": "082b04e8a7c8b26ec6c42f977bfbdc99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found!"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "productPagination": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/product"
                        }
                    },
                    "first_page_url": {
                        "type": "string",
                        "example": "http://api.example.com/api/v1/products?page=1"
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "last_page_url": {
                        "type": "string",
                        "example": "http://api.example.com/api/v1/products?page=10"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string"
                                },
                                "active": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "http://api.example.com/api/v1/products?page=2",
                        "nullable": true
                    },
                    "path": {
                        "type": "string",
                        "example": "http://api.example.com/api/v1/products"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
...
            "storeProduct": {
                "required": [
                    "name",
                    "brand",
                    "price",
                    "quantity",
                    "options",
                    "category_id",
                    "stock",
                    "user_id"
                ],
                "properties": {
                    "name": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "brand": {
                        "description": "Product brand",
                        "type": "string"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "float"
                    },
                    "category_id": {
                        "description": "Category ID",
                        "type": "integer"
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string",
                        "nullable": true
                    },
                    "short_description": {
                        "description": "Short product description",
                        "type": "string",
                        "nullable": true
                    },
                    "stock": {
                        "description": "Stock quantity",
                        "type": "integer",
                        "example": 10
                    },
                    "options": {
                        "description": "Available product options (color, size, etc.)",
                        "type": "object",
                        "example": {
                            "color": [
                                "red",
                                "blue"
                            ],
                            "size": [
                                "small",
                                "medium"
                            ]
                        }
                    },
                    "discount": {
                        "description": "Discount percentage",
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "is_top": {
                        "description": "Is top product flag",
                        "type": "boolean",
                        "nullable": true
                    },
                    "status": {
                        "description": "Product status",
                        "type": "string",
                        "enum": [
                            "available",
                            "out_of_stock",
                            "discontinued"
                        ],
                        "nullable": true
                    },
                    "rating": {
                        "description": "Product rating",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "reviews": {
                        "description": "Number of reviews",
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "updateProduct": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "quantity",
                    "options"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "options": {
                        "description": "Available product options",
                        "type": "array",
                        "items": {
                            "properties": {
                                "color": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "size": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}